+++ client/src/lib/Diff/DiffSelection.svelte
@@ -23,10 +23,25 @@
     TableHeadCell
   } from "flowbite-svelte";
   import { push } from "svelte-spa-router";
-  import { getPublisher, getRelativeTime, request } from "$lib/utils";
+  import { request } from "$lib/utils";
+  import { getPublisher } from "$lib/publisher";
   import ErrorMessage from "$lib/Errors/ErrorMessage.svelte";
   import { getErrorDetails, type ErrorDetails } from "$lib/Errors/error";
   import { onDestroy, onMount } from "svelte";
+
+  const getRelativeTime = (date: Date) => {
+    const now = Date.now();
+    const passedTime = now - date.getTime();
+    if (passedTime < 60000) {
+      return "<1 min ago";
+    } else if (passedTime < 3600000) {
+      return `${Math.floor(passedTime / 60000)} min ago`;
+    } else if (passedTime < 86400000) {
+      return `${Math.floor(passedTime / 3600000)} hours ago`;
+    } else {
+      return `${Math.floor(passedTime / 86400000)} days ago`;
+    }
+  };
 
   $: $appStore.app.diff.docA_ID, getDocuments();
   $: $appStore.app.diff.docB_ID, getDocuments();
